Fix lua FieldInfo sort ordering
Left and right were swapped in the __le and __lt functions, fixed.
Also, since start+length points to the byte *after* the end of the
field, changed the operator in __lt to a <= instead of a <.

This is technically a breaking change, but it does bring the behavior
into line with the documentation.