Remove unnecessary null check in SequenceDialog
SequenceDialog::diagramClicked is checking whether the 'event' argument is
null, but later it dereferences 'event' outside the conditional, so if it
was null it would crash anyway. It doesn't seem possible for the event
argument to actually be null, so this commit removes the redundant check
here and in mouseMoved. I'm also adding an assert to document the non-null
assumption.

Bug found by clang static analyzer.

Fixes #17426.