RTP Analysis：Fix nominal and arrival times calculation
Previously, the nominal and arrival times are calculated based on first packet
in the RTP stream, but there is a corner case: if the stream codec changes in the
middle, e.g., from AMR-WB to AMR, the nominal time will be calculated using the
current codec frequency, and it is not correct and will affect diff and jitter.

This fix will calculate nominal and arrival times based on previous in-sequence
RTP packet.