tcp: Send the header to the tap even with exceptions in options
The TCP tap sends the header information to its listeners, most of
which only need the information in the beginning of the fixed header.
It doesn't get sent until after parsing the TCP options, in order
to add information about SACKs (which most taps don't need, but
at least the TCP stream graph can take advantage of if there.)

Add a cleanup function so that if we hit an exception while parsing
the options, we go ahead and send what we have in the header to the
taps.

Fix #18138.