ANSI IS-637 A: Use tvb_new_octet_aligned for addresses
Use tvb_new_octet_aligned when adding addresses (strings or bytes)
that are not byte aligned. That is not only clearer code, but also
prevents attempting to add unvalidated strings.

Since we're aligning the fields properly, get rid of the extra
fields for the MSB of the first field and LSB of the last field.

Fix #18664